Apply lints and cleanup to Rust code

+++ document-legacy/src/layers/folder_layer.rs
@@ -72,7 +72,7 @@ impl FolderLayer {
 		let mut insert_index = insert_index as i128;
 
 		if insert_index < 0 {
-			insert_index = self.layers.len() as i128 + insert_index as i128 + 1;
+			insert_index = self.layers.len() as i128 + insert_index + 1;
 		}
 
 		if insert_index <= self.layers.len() as i128 && insert_index >= 0 {
